+++ protobuf/lib/src/protobuf/unknown_field_set.dart
@@ -5,6 +5,7 @@
 part of protobuf;
 
 class UnknownFieldSet {
+  static final UnknownFieldSet emptyUnknownFieldSet = UnknownFieldSet().._markReadOnly();
   final Map<int, UnknownFieldSetField> _fields =
       new Map<int, UnknownFieldSetField>();